+++ Challenges/SectionEight/src/main/java/org/example/GasPoweredCar.java
@@ -1,10 +1,12 @@
 package org.example;
 
+import java.math.BigDecimal;
+
 public class GasPoweredCar extends Car {
-    private double avgKmPerLitre;
+    private BigDecimal avgKmPerLitre;
     private int cylinders;
 
-    public GasPoweredCar(String description, double avgKmPerLitre, int cylinders) {
+    public GasPoweredCar(String description, BigDecimal avgKmPerLitre, int cylinders) {
         super(description);
         this.avgKmPerLitre = AutoChecks.checkAvgKmPerLitre(avgKmPerLitre);
         this.cylinders = AutoChecks.checkCylinders(cylinders);
@@ -26,11 +28,11 @@ public class GasPoweredCar extends Car {
         return super.runEngine() + " and is loud given the large amount of horsepower.";
     }
 
-    public double getAvgKmPerLitre() {
+    public BigDecimal getAvgKmPerLitre() {
         return avgKmPerLitre;
     }
 
-    public void setAvgKmPerLitre(double avgKmPerLitre) {
+    public void setAvgKmPerLitre(BigDecimal avgKmPerLitre) {
         this.avgKmPerLitre = AutoChecks.checkAvgKmPerLitre(avgKmPerLitre);
     }
